fix: fix new fam line up logic

+++ src/main/java/org/sarangchurch/growing/v1/feat/newfamily/infra/component/lineinout/NewFamilyLineOutManager.java
@@ -5,7 +5,7 @@ import org.sarangchurch.growing.core.interfaces.common.Events;
 import org.sarangchurch.growing.v1.feat.newfamily.domain.NewFamiliesEmitEvent;
 import org.sarangchurch.growing.v1.feat.newfamily.domain.lineoutnewfamily.LineOutNewFamily;
 import org.sarangchurch.growing.v1.feat.newfamily.domain.newfamily.NewFamily;
-import org.sarangchurch.growing.v1.feat.newfamily.infra.component.NewFamilyPromoteLogManager;
+import org.sarangchurch.growing.v1.feat.newfamily.domain.newfamily.NewFamilyStatus;
 import org.sarangchurch.growing.v1.feat.newfamily.infra.data.lineoutnewfamily.LineOutNewFamilyWriter;
 import org.sarangchurch.growing.v1.feat.newfamily.infra.data.newfamily.NewFamilyFinder;
 import org.sarangchurch.growing.v1.feat.newfamily.infra.data.newfamily.NewFamilyWriter;
@@ -19,8 +19,6 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class NewFamilyLineOutManager {
     private final NewFamilyFinder newFamilyFinder;
-    private final NewFamilyPromoteLogManager newFamilyPromoteLogManager;
-
     private final NewFamilyWriter newFamilyWriter;
     private final LineOutNewFamilyWriter lineOutNewFamilyWriter;
 
@@ -28,8 +26,12 @@ public class NewFamilyLineOutManager {
     public List<LineOutNewFamily> lineOut(List<Long> newFamilyIds) {
         List<NewFamily> newFamilies = newFamilyFinder.findByIdInOrThrow(newFamilyIds);
 
-        // 새가족 라인아웃은 등반 이전에만 가능함
-        newFamilyPromoteLogManager.validateBeforePromotedByNewFamilyIds(newFamilyIds);
+        boolean containsPromoted = newFamilies.stream()
+                .anyMatch(it -> it.statusEquals(NewFamilyStatus.PROMOTED));
+
+        if (containsPromoted) {
+            throw new IllegalStateException("이미 등반된 새가족이 포함되어 있습니다.");
+        }
 
         newFamilyWriter.deleteByIdIn(newFamilyIds);
